requirements.txt
Flask>=2.0.0
yfinance>=0.2.0
pandas>=1.5.0
psutil>=5.8.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pytz>=2021.1
SQLAlchemy>=1.4.0
APScheduler>=3.9.0
psycopg2-binary>=2.9.0